Adjust dataCleaning
The previous process wasn't working properly. Updated with a new data cleaning keeping the operations in Mongosh and the main filterings, but changing the process used for the 'venue' data. Python code to manage 'venue' and vim substitution to obtain integers. Everything is working, but with the using data after the filtering from 2300 elements we have only 600 elements. Try with a bigger dataset.